requirements.txt
streamlit>=1.32
pandas>=2.2
pyarrow>=15
plotly>=5.20
numpy>=1.26
Pillow>=10
